Tried to implement screens

+++ Game/core/src/org/academiadecodigo/hashtronauts/levels/gameObjects/Platform.java
@@ -2,7 +2,7 @@ package org.academiadecodigo.hashtronauts.levels.gameObjects;
 
 public class Platform extends GameObject {
 
-    public Platform(String texture, int posX, int posY, int width, int height){
+    public Platform(String texture, float posX, float posY, int width, int height){
         super(texture, posX, posY, width, height);
     }
 }
